Revised migrations and inflection added

+++ db/migrate/20100310184909_rename_internaluser_to_user.rb
@@ -1,11 +1,41 @@
 class RenameInternaluserToUser < ActiveRecord::Migration
   def self.up
-    rename_column :usercredits, :internalusergive_id, :usergive_id
-    remove_column :usercredits, :usergive_is_internal
+    if table_exists? :usercredits
+      rename_column :usercredits, :internalusergive_id, :usergive_id
+      remove_column :usercredits, :usergive_is_internal
+    else
+      create_table :usercredits do |t|
+        t.integer :internalusergive_id, :externalusergive_id, :month
+        t.integer :year, :null => false
+        t.boolean :usergive_is_internal
+        t.text :other
+      end
+    end
     
-    rename_column :userrefereedpubs, :internaluser_id, :user_id
-    rename_column :userresearchgroups, :internaluser_id, :user_id
-    
+    if table_exists? :userrefereedpubs 
+      if column_exists?(:userrefereedpubs, :internaluser_id)
+        rename_column :userrefereedpubs, :internaluser_id, :user_id
+      else
+        create_table :userrefereedpubs do |t|
+          t.integer :refereedpubs_id, :null => false
+          t.boolean :user_is_internal
+          t.integer :externaluser_id
+          t.integer :internaluser_id
+        end
+      end
+    end
+    if table_exists? :userresearchgroups
+      if column_exists? :userresearchgroups, :internaluser_id
+        rename_column :userresearchgroups, :internaluser_id, :user_id
+      else
+        create_table :userresearchgroups do |t|
+          t.integer :researchgroup_id, null => false
+          t.integer :year, null => false
+          t.boolean :user_is_internal
+          t.integer :externaluser_id, :internaluser_id, :moduser_id
+        end
+      end
+    end
   end
 
   def self.down
